requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
blinker==1.6.2
bottle==0.12.25
branca==0.6.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.3
django-anymail==10.0
django-appconf==1.0.5
django-ckeditor==6.5.1
django-cogwheels==0.3
django-crispy-forms==2.0
django-filer==3.0.1
django-filter==22.1
django-imagekit==4.1.0
django-ipware==5.0.0
django-js-asset==2.0.0
django-modelcluster==6.0
django-multiselectfield==0.1.12
django-multiupload==0.6.1
django-permissionedforms==0.1
django-polymorphic==3.1.0
django-smtp-ssl==1.0
django-storages==1.13.2
django-taggit==3.1.0
django-templated-mail==1.1.1
django-tinymce==3.6.1
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
draftjs-exporter==2.1.7
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
facebook-sdk==3.1.0
Faker==19.2.0
filetype==1.2.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
folium==0.14.0
future==0.18.3
geocoder==1.38.1
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
html5lib==1.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
l18n==2021.3
lxml==4.9.3
MarkupSafe==2.1.3
msgpack==1.0.5
mysqlclient==2.2.0
numpy==1.25.0
oauthlib==3.2.2
openpyxl==3.1.2
pilkit==2.0
Pillow==9.5.0
protobuf==4.23.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
ratelim==0.1.6
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
secure-smtplib==0.1.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
SQLAlchemy==2.0.17
sqlparse==0.4.4
svglib==1.5.1
telepath==0.3
tinycss2==1.2.1
typing_extensions==4.7.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wagtail==5.0.2
wagtailmenus==3.1.8
webencodings==0.5.1
Werkzeug==2.3.6
Willow==1.5
WTForms==3.0.1
